Bring things a bit closer to earlier version
