+++ package.json
@@ -3,7 +3,7 @@
   "scripts": {
     "lint": "eslint *.js src tools",
     "relay-compiler": "relay-compiler --src src --exclude 'data/**' --schema src/data/schema.graphql",
-    "start": "npm run relay-compiler && babel-node index.js",
+    "start": "npm run relay-compiler && babel-node src/server",
     "update-schema": "babel-node tools/updateSchema.js"
   },
   "lint-staged": {
@@ -19,7 +19,6 @@
     "trailingComma": "all"
   },
   "dependencies": {
-    "@artsy/express-reloadable": "^1.4.0",
     "babel-cli": "^6.26.0",
     "babel-core": "^6.26.3",
     "babel-eslint": "^8.2.3",
@@ -45,7 +44,6 @@
     "farce": "^0.2.6",
     "found": "^0.3.11",
     "found-relay": "^0.3.0-alpha.13",
-    "friendly-errors-webpack-plugin": "^1.7.0",
     "graphql": "^0.13.2",
     "graphql-relay": "^0.5.5",
     "husky": "^0.14.3",
@@ -54,18 +52,17 @@
     "lint-staged": "^7.1.2",
     "prettier": "^1.12.1",
     "prop-types": "^15.6.1",
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2",
-    "react-relay": "~1.5.0",
+    "react": "^16.4.0",
+    "react-dom": "^16.4.0",
+    "react-relay": "^1.6.0",
     "react-relay-network-modern": "^2.3.0",
     "react-relay-network-modern-ssr": "1.1.2",
-    "regenerator-runtime": "^0.11.1",
-    "relay-compiler": "~1.5.0",
-    "relay-runtime": "~1.5.0",
+    "relay-compiler": "^1.6.0",
+    "relay-runtime": "^1.6.0",
     "serialize-javascript": "^1.5.0",
-    "todomvc-app-css": "^2.1.1",
+    "todomvc-app-css": "^2.1.2",
     "todomvc-common": "^1.0.5",
-    "webpack": "^4.7.0",
+    "webpack": "^4.8.3",
     "webpack-dev-middleware": "^3.1.3"
   }
 }
